59_выносит дублированную логику наполнения контентов в отдельный класс, провайдит этот класс во вьюхолдер и bottomsheedialog
