+++ app/src/main/java/com/davay/android/extensions/IntExtensions.kt
@@ -10,21 +10,3 @@ fun Int.dpToPx(): Float {
     val metrics = Resources.getSystem().displayMetrics
     return this * (metrics.densityDpi.toFloat() / DisplayMetrics.DENSITY_DEFAULT)
 }
-
-fun Int.formatMovieDuration(context: Context): String {
-    val hours = this / MINUTES_IN_HOUR
-    val remainingMinutes = this % MINUTES_IN_HOUR
-
-    return when {
-        hours > 0 && remainingMinutes > 0 -> context.getString(
-            R.string.select_movies_hours_and_minutes,
-            hours,
-            remainingMinutes
-        )
-
-        hours > 0 -> context.getString(R.string.select_movies_hours, hours)
-        else -> context.getString(R.string.select_movies_minutes, remainingMinutes)
-    }
-}
-
-const val MINUTES_IN_HOUR = 60
