+++ app/src/main/java/com/davay/android/feature/match/presentation/MatchBottomSheetFragment.kt
@@ -7,17 +7,11 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.FrameLayout
-import android.widget.ImageView
-import android.widget.TextView
 import android.widget.Toast
-import coil.load
-import coil.size.Scale
-import coil.transform.RoundedCornersTransformation
-import com.davay.android.R
 import com.davay.android.databinding.FragmentMatchBottomSheetBinding
-import com.davay.android.extensions.formatMovieDuration
 import com.davay.android.feature.selectmovie.MovieDetailsDemo
-import com.google.android.flexbox.FlexboxLayout
+import com.davay.android.utils.MovieDetailsHelper
+import com.davay.android.utils.MovieDetailsHelperImpl
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
@@ -28,6 +22,8 @@ class MatchBottomSheetFragment(
     private var _binding: FragmentMatchBottomSheetBinding? = null
     private val binding: FragmentMatchBottomSheetBinding
         get() = _binding!!
+
+    private val movieDetailsHelper: MovieDetailsHelper = MovieDetailsHelperImpl()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -103,83 +99,13 @@ class MatchBottomSheetFragment(
 
     private fun fillData(data: MovieDetailsDemo) {
         binding.matchMovieCard.apply {
-            setImage(ivSelectMovieCover, data.posterUrl)
-            addGenreList(fblGenreList, data.genres)
+            movieDetailsHelper.setImage(ivSelectMovieCover, data.posterUrl)
+            movieDetailsHelper.addGenreList(fblGenreList, data.genres, requireContext())
             tvFilmTitle.text = data.movieName
             tvOriginalTitle.text = data.alternativeName ?: data.englishName ?: ""
-            tvYearCountryRuntime.text = buildStringYearCountriesRuntime(data)
-            setRateText(tvMarkValue, data.ratingKinopoisk)
+            tvYearCountryRuntime.text =
+                movieDetailsHelper.buildStringYearCountriesRuntime(data, requireContext())
+            movieDetailsHelper.setRateText(tvMarkValue, data.ratingKinopoisk, requireContext())
         }
-    }
-
-    private fun setImage(img: ImageView, url: String?) {
-        img.load(url) {
-            placeholder(com.davai.uikit.R.drawable.placeholder_general_80)
-                .scale(Scale.FIT)
-            error(R.drawable.ic_movie_selection_error_332)
-                .scale(Scale.FIT)
-            transformations(RoundedCornersTransformation())
-                .crossfade(true)
-        }
-    }
-
-    private fun setRateText(tvRate: TextView, ratingKinopoisk: Float?) {
-        ratingKinopoisk?.let {
-            val textColor = if (it >= GOOD_RATE_7) {
-                resources.getColor(com.davai.uikit.R.color.done, null)
-            } else {
-                resources.getColor(com.davai.uikit.R.color.attention, null)
-            }
-            tvRate.apply {
-                text = it.toString()
-                setTextColor(textColor)
-            }
-        }
-    }
-
-    private fun addGenreList(flexBox: FlexboxLayout, genres: List<String>) = with(binding) {
-        flexBox.removeAllViews()
-        genres.forEach {
-            val genreView = LayoutInflater.from(root.context)
-                .inflate(
-                    R.layout.item_genre,
-                    flexBox,
-                    false
-                ) as TextView
-            genreView.text = it
-            flexBox.addView(genreView)
-        }
-    }
-
-    private fun buildStringYearCountriesRuntime(data: MovieDetailsDemo): String {
-        with(data) {
-            val str = StringBuilder()
-            year?.let {
-                str.append(it)
-                str.append(DOT_DELIMETER)
-            }
-            if (countries.isNotEmpty()) {
-                val countryList = countries.take(MAX_COUNTRY_NUMBER)
-                val countriesString =
-                    countryList.joinToString(separator = DOT_DELIMETER)
-                str.append(countriesString)
-                if (countries.size > MAX_COUNTRY_NUMBER) {
-                    str.append(MULTIPOINT)
-                }
-            }
-            movieLengthMin?.let {
-                str.append(DOT_DELIMETER)
-                str.append(movieLengthMin.formatMovieDuration(requireContext()))
-            }
-            return str.toString()
-        }
-    }
-
-    companion object {
-        const val MATCH_KEY = "match_key"
-        const val DOT_DELIMETER = " ∙ "
-        const val MULTIPOINT = "..."
-        const val MAX_COUNTRY_NUMBER = 3
-        const val GOOD_RATE_7 = 7.0f
     }
 }
